src/ams-asset-player/requirements.txt
Flask
freeze
azure-mgmt-resource
azure-mgmt-media
python-dotenv
pandas